+++ ui/page/wallet/transactionListTable/view.jsx
@@ -1,9 +1,9 @@
 // @flow
 import * as MODALS from 'constants/modal_types';
 import React from 'react';
-import TxoListItem from 'component/transactionListTableItem';
 import Spinner from 'component/spinner';
 import LbcSymbol from 'component/common/lbc-symbol';
+import TxoListItem from '../transactionListTableItem';
 
 type Props = {
   emptyMessage: ?string,
